Update Payment to PaymentLeg
Renamed Payment to PaymentLeg to prepare for a later framing where a
Payment is actually a series of legs, each on a different ERC20 supply.

Added a erc20 field to PaymentLeg to support different target supplies.
Added computation of PaymentLeg.id by hashing fields. ID is stored
internally as a bytes32 hash, but external party can give a 0x string
and find a unique PaymentLeg - string is converted to bytes32 to be
compared with internal record.

Added helper Common.newLeg() to generate new PaymentLeg instance, so
that id could be computed internally. Prior to this, PaymentLeg was
being explicitly instantiated by setting id to array index.

Note - netting is still part of PaymentSystem contract, which should
ideally be moved out to a separate contract. Also, netting does not
take into account the erc20 supply which each PaymentLeg targets,
which needs a major update to do netting across multiple supplies,
where each supply can net its legs.

Also need a way to track original payments that are being netted,
so that if there are 'n' payments added to PaymentSystem, should be
possible to net some collection of less than 'n' legs (the choice
of which is outside the scope of the PaymentSystem).