Add INetter & SingleCurrencyNetter
Move netting logic from PaymentSystem to SingleCurrencyNetter, while
adding use of the new interface.

SingleCurrencyNetter has state variables to do the offsetting which
can probably be optimized out -- there is no need to store that info as
state when it is really needed only during an offsetting run and to
return the output of that run to the calling contract.

Also, PaymentSystem doesn't yet invoke the SingleCurrencyNetter, which
needs to be added. Refactored code compiles.