Connect PaymentSystem to Netter
Main change is to invoke the Netter to do offsetting on the payment
system's rawIntentions, to compute nettedIntentions. Minor change
to rename the arrays in PaymentSystem to rawIntentions and
nettedIntentions.

Converting from the Netter's return value in memory to a storage
variable nettedIntentions requires traversing and copying from the
memory array, item by item. TBD if this can be made more efficient.