Set input arg type for INetter interface method
Input arg was previously set as type memory. This can work, but is not
efficient for two reasons: the array is copied to be passed in, bad if
the array is large as it could be in this case; secondly it signals to
the called function that the parameter is read-write. Instead, making
it calldata avoids copying (it is passed by reference), and signals to
the called function that it is read-only -- which is the intended use
in this case. Hence, input arg has been changed to calldata.

The return value must be declared as memory, so no change there. Also,
a dynamically sized return value (like an array in this case) must be
returned _from_ memory by the caller -- this has been clarified in a
comment in INetter so any future implementation is aware of this.