Return payment leg ID as a string
A payment leg ID is stored as a bytes32 hash, since bytes32 vars can
be compared using the "==" operator, which strings cannot be. This has
not changed.

However, earlier the ID of a new payment leg was being returned to the
caller as a bytes32, but bytes32 is not human-readable. With this edit
the ID is returned to the caller as a string that is human-readable.
On subsequent calls to operate on that payment, caller would give the
human-readable string (as before) which the contract converts to a
bytes32 hash before comparing it to IDs in the state storage.