Update Netter method to take netting currency arg
Update INetter method to accept argument of ERC20 address to net by.
Also cleaned up PaymentSystem to use just one NETTER reference.

For now, the Netter implementation expects erc20 arg to be passed in as
address(0), and moreover, just assumes all raw payments are in 1 erc20.
Netter behavior is undefined if payments are in more than one currency.

In upcoming commits, will update to support specifying a single erc20
to do netting by, and to do actual multi-lateral netting.

This commit compiles, but functionality is incomplete!