Add test/basic.js and .gitignore
The 'test' folder was incorrectly in .gitignore, so test code wasn't in
the repo. Fixed .gitignore to remove test/, and also removed .gitignore
from .gitignore so future changes to that file can be tracked.

This adds test/basic.js to the repo. There is also test/Lock.js, which
is a sample test file that goes with the sample Lock.sol contract, both
of which can be ignored.