Add onebank.js test file with related changes
Added new test file onebank.js. It has important updates in its preface
to support reverts and revertsWith Chai constructs with ethers v6. It
also uses loadFixture().  Also, 3 tests related to adding a payment
intention, retriving its PID, and deleting it are added and all pass.

Updates to PaymentSystem to store lastPid as contract state. This
is needed because intentToPay() returns PID of new leg only if called
by another contract. When invoked by remote client, it returns a
transaction receipt, so cannot be used to retrieve and store the PID.
For now, one way that works is to compute the PID in the contract and
store it in contract state. Then use a subsequent 'view' getter API to
read it. Another way would have been for the caller to compute the PID
and specify it in the call to intentToPay(), which would probably be
the better solution (because in a high-volume scenario the lastPid in
the contract could be overwritten before it is read).

Also added PaymentSystem API getPayment() to retrieve via pid. Older
API to retrieve by array index has been deleted, not needed.

Other updates to PaymentSystem to do better logging and some reverts.

Added myver() pure function to MultilateralNetter for easy check of
correct install.