Make localhost the default n/w in hardhat config
Change default network from opencbdc to localhost, to make it easier
to test with local Hardhat node. With this change, every npx hardhat
command like "npx hardhat --network locahost console" can be run as
"npx hardhat console".

Note, since the Hardhat local node (started with npx hardhat node)
starts and displays 20 private keys and associated addresses that
are pre-configured for testing. These do not need to be added in the
hardhat.config.js (like they are for the opencbdc config).

To access the private keys and associated signers in hardhat console:
    > sigs = await ethers.getSigners()
    > sigs.length // returns 20