Add onebank test & edits to use signer to deploy
Added onebank.js test to add multiple payment legs and retrieve/verify.
Test passes.

Also fixed README to clarify that signer #1 is not associated with any
entity. It is the default signer used to deploy libraries and do common
actions not associated with a specific person or bank. Re-assigned keys
to entities such as odd-numbered keys belong to country A entities, and
even-numbered keys to country B entities.

Updated fixture function to use key #1 (the CorrA key) when installing
that bank's ERC20 contract. This isn't needed for the tests to pass,
is a good convention to use a bank's signer when installing its ERC20
contract.