Set endpt of netted payments to the PaymentSystem
Before this change, netted payments had one endpoint as the actual
sender or receiver, and the other endpoint as the ERC20 contract
address. The latter is wrong.

This commit sets the endpoint to the PaymentSystem contract address
and not the ERC20 address. This way, incoming payments to the system
would be paid to the PaymentSystem contract, and outgoing ones would
go from that contract to the respective users.

The netter sets the endpoint now to address(0) instead of the ERC20.
The PaymentSystem checks if an endpoint is zero, and replaces it with
its own address, before storing the netted payments as contract state.

Changed all onebank.js tests related to netting to check for the
new endpoint, all tests pass.