Add onebank test to settle two raw payments
Two payments are A->B $75, and B->A $60, while A and B each have $50
only. This shows that funds are not locked in advance, since if we
tried to lock $75 and $60 on A and B respectively, it would fail due
to insufficient funds.

Instead, the raw payments only require approvals for $75 and $60 to be
recorded, which succeed as approvals don't require funds to be present.
Later, netting happens and when the ERC20 transfers occur, only $15
moves between the accounts, which succeeds as expected.

Test passes, all good. PaymentSystem.undoClear() not implemented or
tested yet.